feat: [#51925] make queues quorum by default
The queue type has been changed from 'classic' to 'quorum'[1]. The
primary motivation is to ensure messages are rejected to the _back_ of
the queue. This allows us to consume as many messages as possible,
instead of blocking on the first unprocessable one.

Even though RabbitMQ queues cannot change their type, but can only be
set on declaration time, the change is not breaking. This is because
queues are migrated on a best-effort basis, i.e. if they are empty, and
thus able to be re-created on startup. Non-empty queues can be deleted
manually using the following commands:
```
docker exec msg_broker rabbitmqctl list_queues
docker exec msg_broker rabbitmqctl delete_queue
```

[1] https://www.rabbitmq.com/quorum-queues.html