feat: [#59372] add caching to from_context
This commit adds caching to the `from_context` dependency injection
helper, the motivation behind this change is easing testing when using
Annotated types, depending on `from_context` functions.

For example, imagine we have a dependency alike this:
```python
GQLClient = Annotated[_GQLClient, Depends(from_context("graphql_client"))]
```

Now imagine that we wish to replace the GraphQL client with an `AsyncMock`
in our test-suite, to do this we will use `app.dependency_overrides`:
```python
    mock = AsyncMock()
    def mocker():
      return mock
    app.dependency_overrides[from_context("graphql_client")] = mocker
```
However this example depends on multiple calls to `from_context`
returning the **exact same** inner-function every time, as otherwise the
dependency override will not hit.

Caching the `from_context` function ensures that we always get the
**exact same** inner-function as a result.